@@ -4,12 +4,83 @@ import android.os.Bundle
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
+import androidx.fragment.app.Fragment
 import com.spogss.scrumble.R
+import com.spogss.scrumble.data.Project
+import com.spogss.scrumble.data.Sprint
+import com.spogss.scrumble.data.Task
+import com.spogss.scrumble.data.User
+import com.spogss.scrumble.enums.UserStoryState
+import java.util.*
 
 
-class MyTasksFragment: BoardViewFragment() {
+class MyTasksFragment: Fragment() {
+    //TODO: remove when webservice call is possible
+    private var users = mutableListOf<User>()
+    private var projects = mutableListOf<Project>()
+    private var sprints = mutableListOf<Sprint>()
+    private var userStories = mutableListOf<Task>()
+
     override fun onCreateView(inflater: LayoutInflater, container: ViewGroup?, savedInstanceState: Bundle?): View? {
-        layoutRes = R.layout.fragment_my_tasks
-        return super.onCreateView(inflater, container, savedInstanceState)
+        //TODO: remove when webservice call is possible
+        createTestData()
+
+        return inflater.inflate(R.layout.fragment_my_tasks, container, false)
+    }
+
+    override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
+        super.onViewCreated(view, savedInstanceState)
+    }
+
+
+    //TODO: remove when webservice call is possible
+    private fun createTestData() {
+        users.clear()
+        projects.clear()
+        sprints.clear()
+        userStories.clear()
+
+        val sam = User(0, "Sam", "entersamspassworthere")
+        val pauli = User(1, "Pauli", "phaulson")
+        val simon = User(2, "Simon", "hamsterbacke")
+        val webi = User(3, "Webi", "livepoolftw")
+        val savan = User(4, "Savan", "chelseafcgothistory")
+        users.add(sam)
+        users.add(pauli)
+        users.add(simon)
+        users.add(webi)
+        users.add(savan)
+
+        val scrumble = Project(0, "Scrumble", sam, mutableListOf(pauli, simon, webi), null)
+        val footballsimulator = Project(1, "Football Simulator", savan, mutableListOf(webi), null)
+        projects.add(scrumble)
+        projects.add(footballsimulator)
+
+        val scrumble1 = Sprint(0, 1, Date(2018, 10, 3), Date(2018, 10, 12), 0)
+        val scrumble2 = Sprint(1, 2, Date(2018, 10, 12), Date(2018, 11, 1), 0)
+        val scrumble3 = Sprint(2, 3, Date(2018, 11, 3), Date(2018, 11, 22), 0)
+        val footballsimulator1 = Sprint(3, 1, Date(2018, 10, 10), Date(2018, 10, 30), 1)
+        val footballsimulator2 = Sprint(4, 2, Date(2018, 11, 1), Date(2018, 11, 15), 1)
+        sprints.add(scrumble1)
+        sprints.add(scrumble2)
+        sprints.add(scrumble3)
+        sprints.add(footballsimulator1)
+        sprints.add(footballsimulator2)
+
+        val scrumbleGUI = Task(0, pauli, sam, "GUI Programmieren", "Hierbei handelt es sich um die Android GUI",
+                0, UserStoryState.IN_PROGRESS, 0, 0, 0)
+        val scrumbleDB = Task(1, webi, simon, "Datenbank aufsetzten", "Inklusive Trigger, Sequenzen usw...",
+                0, UserStoryState.TO_VERIFY, 1, 0, 0)
+        val scrumbleWebservice = Task(2, webi, pauli, "Webservice implementieren", "Der Webservice connected Client mit Datenbank",
+                0, UserStoryState.SPRINT_BACKLOG, 0, null, 0)
+        val scrumbleTesting = Task(3, simon, sam, "Testen", "Sollen wir Unit-Tests machen? Keine Ahnung",
+                0, UserStoryState.SPRINT_BACKLOG, 0, null, 0)
+        val scrumblePlaning = Task(4, sam, pauli, "Planung", "Geplant wird immer, weil wir mit Scrum arbeiten, lol. Hier steht auch noch mehr Text",
+                0, UserStoryState.IN_PROGRESS, 2, 0, 0)
+        userStories.add(scrumbleGUI)
+        userStories.add(scrumbleDB)
+        userStories.add(scrumbleWebservice)
+        userStories.add(scrumbleTesting)
+        userStories.add(scrumblePlaning)
     }
 }
